requirements.txt
tensorflow-cpu==2.10.0
numpy>=1.23.0
scikit-learn>=1.3.0
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
matplotlib>=3.7.0
